vendors back

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -19,6 +20,9 @@ class VendorProductImageGalleryController extends Controller
 
     {
         $product = Product::findOrFail($request->product);
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
         return $dataTable->render('vendor.product.image-gallery.index', compact('product'));
     }
 
@@ -82,6 +86,11 @@ class VendorProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $productImage = ProductImageGallery::findOrFail($id);
+
+        if ($productImage->product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+
         $this->deleteImage($productImage->image);
         $productImage->delete();
